Add files via upload
successfully creating my Netflix clone! While you acknowledge that it may not be perfect, My dedication and effort shine through in every aspect of its development. Despite any imperfections, your commitment to delivering the best possible product is evident. My Netflix clone may not be flawless, but it represents a significant achievement and a testament to my hard work and determination. With continued effort and refinement, it has the potential to become even better over time